Hack with serial numbers
The challenge is to issue passes and support push notifications for
updates with constraints on storage of device library identifiers and
push tokens.

Requests to the endpoint that retrieves a list of pass serial numbers
by device library identifier doesn't include any user identifying
information. This is a challenge when device library identifiers and
push tokens can only be retrieved by a user ID or username.

This hack works if users only need one pass at a time from the service.
Passes are all issued with the same serial number, but requests for pass
updates contain the authorizationToken, in the Authorization request
header, and this can be used to extract the username and a password.
These credentials can then be used to authenticate the request and to
retrieve the correct pass details for that user.